+++ R/TreeNumber.R
@@ -81,7 +81,9 @@
 #' Note that the hyperexponential nature of tree space means that there are &gt;
 #' 2^30 unique 12-tip trees.  As integers &gt; 2^31 are not supported by R,
 #' numbers representing larger trees are represented internally as a vector of
-#' nine-digit integer 'chunks'.
+#' nine-digit integer 'chunks' and passed to the underlying C code, where they
+#' are combined into a single 64-bit integer.  This allows trees with up to
+#' 42 tips to be accommodated.
 #'
 #' @param x Integer identifying the tree (see details).
 #' @param nTip Integer specifying number of tips in the tree.